+++ mosaic-terminal/src/jvmMain/kotlin/com/jakewharton/mosaic/terminal/Tty.kt
@@ -1,0 +1,66 @@
+package com.jakewharton.mosaic.terminal
+
+import java.io.IOException
+import java.nio.file.Files
+import java.nio.file.Path
+import java.nio.file.StandardCopyOption.REPLACE_EXISTING
+import java.util.Locale.US
+
+public actual object Tty {
+	init {
+		loadNativeLibrary("mosaic")
+	}
+
+	public actual fun enableRawMode(): AutoCloseable {
+		val savedConfig = enterRawMode()
+		if (savedConfig == 0L) throw OutOfMemoryError()
+		return RawMode(savedConfig)
+	}
+
+	private class RawMode(
+		private val savedPtr: Long,
+	) : AutoCloseable {
+		override fun close() {
+			val error = exitRawMode(savedPtr)
+			check(error == 0) { "Unable to exit raw mode: $error" }
+		}
+	}
+
+	@JvmStatic
+	private external fun enterRawMode(): Long
+
+	@JvmStatic
+	private external fun exitRawMode(savedConfig: Long): Int
+
+	@Suppress(
+		// Only loading from our own JAR contents.
+		"UnsafeDynamicallyLoadedCode",
+		// Preserving copy/paste!
+		"SameParameterValue",
+	)
+	private fun loadNativeLibrary(name: String) {
+		val osName = System.getProperty("os.name").lowercase(US)
+		val osArch = System.getProperty("os.arch").lowercase(US)
+		val nativeLibraryJarPath = "/jni/$osArch/" + when {
+			"linux" in osName -> "lib$name.so"
+			"mac" in osName -> "lib$name.dylib"
+			"windows" in osName -> "$name.dll"
+			else -> throw IllegalStateException("Unsupported OS: $osName $osArch")
+		}
+		val nativeLibraryUrl = Tty::class.java.getResource(nativeLibraryJarPath)
+			?: throw IllegalStateException("Unable to read $nativeLibraryJarPath from JAR")
+		val nativeLibraryFile: Path
+		try {
+			nativeLibraryFile = Files.createTempFile(name, null)
+
+			// File-based deleteOnExit() uses a special internal shutdown hook that always runs last.
+			nativeLibraryFile.toFile().deleteOnExit()
+			nativeLibraryUrl.openStream().use { nativeLibrary ->
+				Files.copy(nativeLibrary, nativeLibraryFile, REPLACE_EXISTING)
+			}
+		} catch (e: IOException) {
+			throw RuntimeException("Unable to extract native library from JAR", e)
+		}
+		System.load(nativeLibraryFile.toAbsolutePath().toString())
+	}
+}