+++ mosaic-terminal/src/nativeMain/kotlin/com/jakewharton/mosaic/terminal/Tty.kt
@@ -1,0 +1,25 @@
+package com.jakewharton.mosaic.terminal
+
+import kotlinx.cinterop.CPointer
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.cinterop.useContents
+
+@OptIn(ExperimentalForeignApi::class)
+public actual object Tty {
+	public actual fun enableRawMode(): AutoCloseable {
+		val savedConfig = enterRawMode().useContents {
+			check(error == 0U) { "Unable to enable raw mode: $error" }
+			saved ?: throw OutOfMemoryError()
+		}
+		return RawMode(savedConfig)
+	}
+
+	private class RawMode(
+		private val savedConfig: CPointer<rawModeConfig>,
+	) : AutoCloseable {
+		override fun close() {
+			val error = exitRawMode(savedConfig)
+			check(error.toInt() == 0) { "Unable to exit raw mode: $error" }
+		}
+	}
+}