+++ samples/rrtop/src/main/kotlin/example/RrtopApp.kt
@@ -2,13 +2,14 @@ package example
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import com.jakewharton.mosaic.LocalTerminal
+import com.jakewharton.mosaic.layout.KeyEvent
 import com.jakewharton.mosaic.layout.background
 import com.jakewharton.mosaic.layout.fillMaxWidth
 import com.jakewharton.mosaic.layout.height
+import com.jakewharton.mosaic.layout.onKeyEvent
 import com.jakewharton.mosaic.layout.padding
 import com.jakewharton.mosaic.layout.width
 import com.jakewharton.mosaic.modifier.Modifier
@@ -18,17 +19,11 @@ import com.jakewharton.mosaic.text.withStyle
 import com.jakewharton.mosaic.ui.Alignment
 import com.jakewharton.mosaic.ui.Box
 import com.jakewharton.mosaic.ui.Text
-import example.common.Key
-import example.common.key
-import example.common.keyEventFlow
-import example.common.utf16CodePoint
 import example.screens.CmdScreen
 import example.screens.MainScreen
 import example.screens.RawScreen
 import example.screens.SlowScreen
 import example.screens.StatScreen
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 
 @Composable
 fun RrtopApp(rrtopViewModel: RrtopViewModel, colorsPalette: RrtopColorsPalette) {
@@ -39,7 +34,21 @@ fun RrtopApp(rrtopViewModel: RrtopViewModel, colorsPalette: RrtopColorsPalette) 
 			modifier = Modifier
 				.width(terminal.size.width)
 				.height(terminal.size.height - 1) // subtraction of one is necessary, because there is a line with a cursor at the bottom, which moves up all the content
-				.background(LocalRrtopColorsPalette.current.mainBg),
+				.background(LocalRrtopColorsPalette.current.mainBg)
+				.onKeyEvent {
+					when (it) {
+						KeyEvent("ArrowLeft") -> rrtopViewModel.onArrowLeftPress()
+						KeyEvent("ArrowRight") -> rrtopViewModel.onArrowRightPress()
+						KeyEvent("Tab") -> rrtopViewModel.onTabPress()
+
+						KeyEvent("q"),
+						KeyEvent("Q"),
+						-> rrtopViewModel.onQPress()
+
+						else -> return@onKeyEvent false
+					}
+					true
+				},
 		) {
 			when (rrtopUiState.currentScreen) {
 				RrtopUiState.Screen.Main -> MainScreen(
@@ -71,23 +80,6 @@ fun RrtopApp(rrtopViewModel: RrtopViewModel, colorsPalette: RrtopColorsPalette) 
 					.padding(horizontal = 1)
 					.align(Alignment.BottomCenter),
 			)
-		}
-	}
-
-	LaunchedEffect(Unit) {
-		withContext(Dispatchers.IO) {
-			keyEventFlow.collect { keyEvent ->
-				when (keyEvent.key) {
-					Key.DirectionLeft -> rrtopViewModel.onArrowLeftPress()
-					Key.DirectionRight -> rrtopViewModel.onArrowRightPress()
-					Key.Tab -> rrtopViewModel.onTabPress()
-					Key.Type -> {
-						if (keyEvent.utf16CodePoint == 'q'.code || keyEvent.utf16CodePoint == 'Q'.code) {
-							rrtopViewModel.onQPress()
-						}
-					}
-				}
-			}
 		}
 	}
 }